Restructured allnotes layout
