+++ src/app/(frontend)/allnotes/page.tsx
@@ -7,5 +7,5 @@ export default async function AllNotesPage() {
 
   if (!user) return <UnauthenticatedPage />;
 
-  return <p>Select a note to view</p>;
+  return null;
 }
